requirements.txt
discord.py
aiohttp>=3.12.14
beautifulsoup4>=4.13.4
brotli>=1.1.0
psutil>=7.0.0
python-dotenv>=1.1.1
trafilatura>=2.0.0
flask>=3.0.0
